backend/requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.1.0

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0

# Jupyter Notebook Processing
nbformat==5.9.2
nbconvert==7.12.0
jupyter-client==8.6.0

# AI/ML Dependencies
scikit-learn==1.3.2
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
nltk==3.8.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# File Handling
python-magic==0.4.27
python-magic-bin==0.4.14

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Database
pymongo==4.6.0
motor==3.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
asyncio==3.4.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
waitress==2.1.2
